refactor: extract shared capture retitle, enrichment, and clip helpers
- retitleDailyEntry() in capture-note.ts centralizes the verbatim-match
  daily link rewrite (and its $-literal replacer) for both the drain's
  dedup refresh and the enrichment retitle, with direct unit tests.
- generateEnrichment() pulls the AI leg (screenshot read + provider
  call + refusal fallback) out of the reconcile loop.
- PageDescription renamed to PageEnrichment now that it carries a title.
- clipAtWordBoundary() in ai/text.ts replaces the duplicated clipTitle
  in audio-memo-title.ts and describe-page.ts.
- capture.test.ts split into capture-identity/-drain/-enrichment test
  files over a shared capture-harness.ts in-memory fixture.

No behavior change; 87 tests across the touched suites pass.
